% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_profiles.R
\name{current_profiles}
\alias{current_profiles}
\title{Calculate current profiles}
\usage{
current_profiles(IMU)
}
\arguments{
\item{IMU}{A dataframe with raw IMU and eular angle data.}
}
\value{
A dataframe with associated current profile data.
}
\description{
This function calculates current profiles based on IMU data.
}
