% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eular_angles_true.R
\name{eular_angles_true}
\alias{eular_angles_true}
\title{Calculate eular angles}
\usage{
eular_angles_true(IMU)
}
\arguments{
\item{IMU}{A data frame with raw IMU data}
}
\value{
A dataframe with associated eular angles.
}
\description{
This function calculates eular angles from IMU data and
returns a dataframe with associated eular angles.
}
