% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_func.R
\name{model_func}
\alias{model_func}
\title{Calculate poly model predicted values for data}
\usage{
model_func(model_x, x)
}
\arguments{
\item{model_x}{model object}

\item{x}{input data}
}
\value{
A dataframe with model predicted values
}
\description{
This function applied polynomial model to input data and
returns a vector with the associated predicted values.
}
