% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_TCM.R
\name{read_TCM}
\alias{read_TCM}
\title{Read TCM data files}
\usage{
read_TCM(TCM_data_dir, data_file_name, burst_rate)
}
\arguments{
\item{TCM_data_dir}{Path to the input file}

\item{data_file_name}{full name of .csv file}

\item{burst_rate}{burst interval used during data collection}
}
\value{
A list of data from binary files
}
\description{
This function reads TCM data files and exports .csv files
with sensor data and current profile data appended.
}
